refactor(as-4806 refactored code in task-list.js): as-4806 refactored code in task-list.js
